rename the package of the ConcurrentReferenceHashMap
