+++ src/main/java/com/antkorwin/xsync/XMutexFactoryImpl.java
@@ -3,7 +3,7 @@ package com.antkorwin.xsync;
 
 import java.util.concurrent.ConcurrentMap;
 
-import org.hibernate.validator.internal.util.ConcurrentReferenceHashMap;
+import com.antkorwin.xsync.douglea.ConcurrentReferenceHashMap;
 
 
 /**
