+++ src/test/java/com/antkorwin/xsync/XMutexFactoryImplTest.java
@@ -10,7 +10,7 @@ import java.util.stream.IntStream;
 
 import com.antkorwin.commonutils.concurrent.ConcurrentSet;
 import com.antkorwin.commonutils.gc.GcUtils;
-import org.hibernate.validator.internal.util.ConcurrentReferenceHashMap;
+import com.antkorwin.xsync.douglea.ConcurrentReferenceHashMap;
 import org.junit.Test;
 
 import static java.util.stream.Collectors.toList;
